requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-community
python-dotenv
requests 
pandas 